+++ src/util.rs
@@ -378,6 +378,7 @@ pub fn find_venv_info(
     cfg_vers: &Version,
     pypackages_dir: &Path,
     pyflow_dir: &Path,
+    cache_dir: &Path,
 ) -> (PathBuf, Version) {
     let venvs = find_venvs(&pypackages_dir);
     // The version's explicitly specified; check if an environment for that version
@@ -390,7 +391,7 @@ pub fn find_venv_info(
     let py_vers;
     match compatible_venvs.len() {
         0 => {
-            let vers = py_versions::create_venv(&cfg_vers, pypackages_dir, pyflow_dir);
+            let vers = py_versions::create_venv(&cfg_vers, pypackages_dir, pyflow_dir, cache_dir);
             vers_path = pypackages_dir.join(&format!("{}.{}", vers.major, vers.minor));
             py_vers = Version::new_short(vers.major, vers.minor); // Don't include patch.
         }